- used BigDecimal for more accurate calculations instead of int

+++ src/main/java/com/brokerage/api/dto/DepositRequest.java
@@ -1,6 +1,7 @@
 package com.brokerage.api.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record DepositRequest(UUID customerId, int amount) {
+public record DepositRequest(UUID customerId, BigDecimal amount) {
 }
